refactor: reorganize my-page feature structure (#145)
