--- src/features/my-page/components/index.ts
@@ -1,6 +1,0 @@
-export { MyPageGoalSection } from './goal/MyPageGoalSection'
-export { MyPageHeatmap } from './MyPageHeatmap'
-export {
-  MyPageStatsSummary,
-  type MyPageStatsSummaryItem,
-} from './MyPageStatsSummary'
