+++ src/features/my-page/heatmap/index.ts
@@ -1,0 +1,1 @@
+export { MyPageHeatmap } from './MyPageHeatmap'
